fix(renderer): export parse condition function
