+++ packages/react-form-renderer/src/files/parse-condition.d.ts
@@ -1,0 +1,6 @@
+import { AnyObject } from "./common";
+import { ConditionDefinition } from "./condition";
+
+export type ParseCondition = (condition: ConditionDefinition, values: AnyObject) => void;
+declare const parseCondition: ParseCondition
+export default parseCondition;
